+++ DiceBot/Mersenne.cs
@@ -189,37 +189,37 @@ namespace DiceBot
         }
         #endregion
 
-        #region "Get type'double' number"
+        #region "Get type'decimal' number"
         /* generates a random number on [0,1]-real-Interval */
-        public double genrand_real1()
+        public decimal genrand_real1()
         {
-            return genrand_Int32() * (1.0 / 4294967295.0);
+            return genrand_Int32() * (1.0m / 4294967295.0m);
             /* divided by 2^32-1 */
         }
 
         /* generates a random number on [0,1)-real-Interval */
-        public double genrand_real2()
+        public decimal genrand_real2()
         {
-            return genrand_Int32() * (1.0 / 4294967296.0);
+            return genrand_Int32() * (1.0m / 4294967296.0m);
             /* divided by 2^32 */
         }
-        public double genrand_real22()
+        public decimal genrand_real22()
         {
             return genrand_Int32() * (1 / 4294967296);
             /* divided by 2^32 */
         }
         /* generates a random number on (0,1)-real-Interval */
-        public double genrand_real3()
+        public decimal genrand_real3()
         {
-            return (genrand_Int32() + 0.5) * (1.0 / 4294967296.0);
+            return (genrand_Int32() + 0.5m) * (1.0m / 4294967296.0m);
             /* divided by 2^32 */
         }
 
         /* generates a random number on [0,1) with 53-bit resolution*/
-        public double genrand_res53()
+        public decimal genrand_res53()
         {
             UInt32 a = genrand_Int32() >> 5, b = genrand_Int32() >> 6;
-            return (a * 67108864.0 + b) * (1.0 / 9007199254740992.0);
+            return (a * 67108864.0m + b) * (1.0m / 9007199254740992.0m);
         }
         /* These real versions are due to Isaku Wada, 2002/01/09 added */
         #endregion
